+++ MonacoEditorComponent/ts-helpermethods/registerColorProvider.ts
@@ -1,17 +1,17 @@
 ﻿///<reference path="../monaco-editor/monaco.d.ts" />
-declare var Parent: ParentAccessor;
+declare var Accessor: ParentAccessor;
 
 const registerColorProvider = function (languageId) {
     return monaco.languages.registerColorProvider(languageId, {
         provideColorPresentations: function (model, colorInfo, token) {
-            return Parent.callEvent("ProvideColorPresentations" + languageId, [JSON.stringify(colorInfo)]).then(result => {
+            return Accessor.callEvent("ProvideColorPresentations" + languageId, [JSON.stringify(colorInfo)]).then(result => {
                 if (result) {
                     return JSON.parse(result);
                 }
             });
         },
         provideDocumentColors: function (model, token) {
-            return Parent.callEvent("ProvideDocumentColors" + languageId, []).then(result => {
+            return Accessor.callEvent("ProvideDocumentColors" + languageId, []).then(result => {
                 if (result) {
                     return JSON.parse(result);
                 }