+++ MonacoEditorComponent/ts-helpermethods/registerCompletionItemProvider.ts
@@ -1,11 +1,11 @@
 ﻿///<reference path="../monaco-editor/monaco.d.ts" />
-declare var Parent: ParentAccessor;
+declare var Accessor: ParentAccessor;
 
 const registerCompletionItemProvider = function (languageId, characters) {
     return monaco.languages.registerCompletionItemProvider(languageId, {
         triggerCharacters: characters,
         provideCompletionItems: function (model, position, context, token) {
-            return Parent.callEvent("CompletionItemProvider" + languageId, [JSON.stringify(position), JSON.stringify(context)]).then(result => {
+            return Accessor.callEvent("CompletionItemProvider" + languageId, [JSON.stringify(position), JSON.stringify(context)]).then(result => {
                 if (result) {
                     const list: monaco.languages.CompletionList = JSON.parse(result);
 
@@ -17,7 +17,7 @@ const registerCompletionItemProvider = function (languageId, characters) {
             });
         },
         resolveCompletionItem: function (item, token) {
-            return Parent.callEvent("CompletionItemRequested" + languageId, [JSON.stringify(item)]).then(result => {
+            return Accessor.callEvent("CompletionItemRequested" + languageId, [JSON.stringify(item)]).then(result => {
                 if (result) {
                     return JSON.parse(result);
                 }